Back to f64

+++ src/vector3.rs
@@ -9,15 +9,15 @@ use serde_wasm_bindgen::from_value;
 #[wasm_bindgen]
 #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
 pub struct Vector3 {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 #[allow(dead_code)]
 #[wasm_bindgen]
 impl Vector3 {
-    pub fn new(x: f32, y: f32, z: f32) -> Vector3 {
+    pub fn new(x: f64, y: f64, z: f64) -> Vector3 {
         Vector3 { x, y, z }
     }
 
@@ -31,25 +31,25 @@ impl Vector3 {
 
     pub fn from_u32(x: u32, y: u32, z: u32) -> Vector3 {
         Vector3 {
-            x: x as f32,
-            y: y as f32,
-            z: z as f32,
+            x: x as f64,
+            y: y as f64,
+            z: z as f64,
         }
     }
 
     pub fn from_i32(x: i32, y: i32, z: i32) -> Vector3 {
         Vector3 {
-            x: x as f32,
-            y: y as f32,
-            z: z as f32,
+            x: x as f64,
+            y: y as f64,
+            z: z as f64,
         }
     }
 
     pub fn from_f32(x: f32, y: f32, z: f32) -> Vector3 {
         Vector3 {
-            x: x,
-            y: y,
-            z: z
+            x: x as f64,
+            y: y as f64,
+            z: z as f64
         }
     }
 }
@@ -96,10 +96,10 @@ impl Mul<f32> for Vector3 {
     }
 }
 
-impl Div<f32> for Vector3 {
+impl Div<f64> for Vector3 {
     type Output = Vector3;
 
-    fn div(self, scalar: f32) -> Self::Output {
+    fn div(self, scalar: f64) -> Self::Output {
         Vector3 {
             x: self.x / scalar,
             y: self.y / scalar,
@@ -116,11 +116,11 @@ impl PartialEq for Vector3 {
 
 #[allow(dead_code)]
 impl Vector3 {
-    pub fn magnitude(&self) -> f32 {
+    pub fn magnitude(&self) -> f64 {
         (self.x.powi(2) + self.y.powi(2) + self.z.powi(2)).sqrt()
     }
 
-    pub fn dot(&self, other: &Vector3) -> f32 {
+    pub fn dot(&self, other: &Vector3) -> f64 {
         self.x * other.x + self.y * other.y + self.z * other.z
     }
 
@@ -132,11 +132,11 @@ impl Vector3 {
         }
     }
 
-    pub fn angle(&self, other: Vector3) -> f32 {
+    pub fn angle(&self, other: Vector3) -> f64 {
         (self.dot(&other) / (self.magnitude() * other.magnitude())).acos()
     }
 
-    pub fn angle_deg(&self, other: Vector3) -> f32 {
+    pub fn angle_deg(&self, other: Vector3) -> f64 {
         (self.dot(&other) / (self.magnitude() * other.magnitude())).acos() * (180.0 / PI)
     }
